refactor(activity-log): rename ActivityLog page to ActivityLogs
