+++ frontend/src/services/acitivity-log/activity-log.api.ts
@@ -1,7 +1,7 @@
 import { OrderDirection } from '../../common/enums/order-direction.enum';
 import { IActivityLogListItem } from '../../common/interfaces/activity-log.interface';
 import { IPaginatedEntity } from '../../common/interfaces/paginated-entity.interface';
-import { ActivityLogTabs } from '../../pages/ActivityLog';
+import { ActivityLogTabs } from '../../pages/ActivityLogs';
 import API from '../api';
 
 interface IPaginatedActivityLog extends IPaginatedEntity<IActivityLogListItem> {
