+++ frontend/src/services/acitivity-log/activity-log.service.ts
@@ -3,7 +3,7 @@ import { useQuery } from 'react-query';
 import { OrderDirection } from '../../common/enums/order-direction.enum';
 import { ACTIVITY_LOG_ERRORS } from '../../common/errors/entities/activity-log.errors';
 import { IBusinessException } from '../../common/interfaces/business-exception.interface';
-import { ActivityLogTabs } from '../../pages/ActivityLog';
+import { ActivityLogTabs } from '../../pages/ActivityLogs';
 import { getActivityLogs } from './activity-log.api';
 
 export const useActivityLogsQuery = (
